Support for enable password configuration
When enable password is configured in device, ML2 plugin
will not be able to configure the device via SSH or TELNET.

Now this is supported by reading the response of each command
and identifying the prompt.

Modified tox.ini file to get the test cases to execute.

Commented Vyatta test code as it fails unit test.

Change-Id: I3344f63b2f076809046fdbe92b9c9227bc0f1e1c
Closes-Bug: 1481161