Final decomposition from neutron tree
All the vendor code is scheduled to be removed from the neutron
tree (see dependent patch). This patch moves the models, config
files, and entry points into the vendor repo so that it is
independent.

This patch also adds machinery to get Zuul "Depends-On" to work;
kudos to dougwig's c61f920f2a54fd829ee37484b3743ab77b909847 to
make it possible.

Depends-on: Ia366ec28d0ae9fd7f9fe41c9a82d5d44971d27d8
Change-Id: I023284f760fa40b7ad06736f529688394ad27e2c